Use rounding for display
Internally, even if a non-integer is used for the value of matched gems, it seems to round it. However, when displayed, it will include the fractional portion of the match before rounding. This gets strange as it also doesn't include a decimal place making a match of 3 * 1.5 = 4.5 display as "45" instead of "4.5" or even "4". Rounding here seems to fix that.